fixing comments

+++ ProseTutorial/WitnessFunctions.cs
@@ -67,7 +67,7 @@ namespace ProseTutorial {
         }
 
         /// <summary>
-        /// This witness function should deduce the spec for k given the spect for AbsPos     
+        /// This witness function should deduce the spec for k given the spec for AbsPos     
         /// </summary>
         /// <param name="rule"></param>
         /// <param name="spec"></param>
@@ -76,9 +76,6 @@ namespace ProseTutorial {
         /// representative spec type in PROSE – DisjunctiveExamplesSpec.</returns>
         [WitnessFunction(nameof(Semantics.AbsPos), 1)]
         public DisjunctiveExamplesSpec WitnessK(GrammarRule rule, DisjunctiveExamplesSpec spec) {
-
-            //the spec on k for each input state will have type IEnumerable<object> since we will have 
-            //more than one possible output
             var kExamples = new Dictionary<State, IEnumerable<object>>();
             foreach (var example in spec.DisjunctiveExamples) {
                 State inputState = example.Key;
@@ -87,10 +84,10 @@ namespace ProseTutorial {
                 var positions = new List<int>();
                 foreach (int pos in example.Value) {
                     //the positive spec for k
-                    positions.Add((int)pos + 1);
+                    positions.Add(pos + 1);
                     //TODO add the negative spec for k 
                     //uncomment the next statement and replace X by the expression that should return the negative spec
-                    //positions.Add((int)pos - X);
+                    //positions.Add(X);
                 }
                 if (positions.Count == 0) return null;
                 kExamples[inputState] = positions.Cast<object>();
